+++ src/simmate/calculators/clease/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class CleaseConfig(AppConfig):
+    name = "simmate.calculators.clease"