+++ src/simmate/calculators/vasp/workflows/diffusion/neb_single_path_base.py
@@ -6,6 +6,7 @@ from simmate.calculators.vasp.workflows.diffusion.utilities import (
     get_migration_images_from_endpoints,
 )
 from simmate.toolkit.diffusion import MigrationHop
+from simmate.toolkit.diffusion.utilities import clean_start_end_images
 from simmate.workflow_engine import Workflow
 
 
@@ -33,7 +34,7 @@ class SinglePathWorkflow(Workflow):
     """
 
     endpoint_relaxation_workflow: Workflow = None
-
+    endpoint_energy_workflow: Workflow = None
     from_images_workflow: Workflow = None
 
     # Oddly enough, the from_images_workflow and this workflow share a table
@@ -50,6 +51,7 @@ class SinglePathWorkflow(Workflow):
         command: str = None,
         diffusion_analysis_id: int = None,
         is_restart: bool = False,
+        relax_endpoints: bool = True,
         # parameters for supercell and image generation
         nimages: int = 5,
         min_atoms: int = 80,
@@ -66,29 +68,58 @@ class SinglePathWorkflow(Workflow):
             vac_mode=True,
         )
 
-        # Relax the starting supercell structure
-        endpoint_start_state = cls.endpoint_relaxation_workflow.run(
+        # BUG-CHECK to ensure sites are in proper order
+        # Consider moving this check into the get_sc_structures method.
+        supercell_start, supercell_end = clean_start_end_images(
+            supercell_start,
+            supercell_end,
+        )
+        ## end bugcheck
+
+        if relax_endpoints:
+            # Relax the starting supercell structure
+            endpoint_start_state = cls.endpoint_relaxation_workflow.run(
+                structure=supercell_start,
+                command=command,  # subcommands["command_supercell"]
+                directory=directory
+                / f"{cls.endpoint_relaxation_workflow.name_full}.start",
+                is_restart=is_restart,
+            )
+
+            # Relax the ending supercell structure
+            endpoint_end_state = cls.endpoint_relaxation_workflow.run(
+                structure=supercell_end,
+                command=command,  # subcommands["command_supercell"]
+                directory=directory
+                / f"{cls.endpoint_relaxation_workflow.name_full}.end",
+                is_restart=is_restart,
+            )
+
+            # wait for the endpoint relaxations to finish and use them to
+            # update the structures we are using
+            supercell_start = endpoint_start_state.result()
+            supercell_end = endpoint_end_state.result()
+
+        # Run static-energy calculations for the endpoints
+        start_energy = cls.endpoint_energy_workflow.run(
             structure=supercell_start,
             command=command,  # subcommands["command_supercell"]
-            directory=directory / f"{cls.endpoint_relaxation_workflow.name_full}.start",
+            directory=directory / f"{cls.endpoint_energy_workflow.name_full}.start",
             is_restart=is_restart,
         )
-
-        # Relax the ending supercell structure
-        endpoint_end_state = cls.endpoint_relaxation_workflow.run(
+        end_energy = cls.endpoint_energy_workflow.run(
             structure=supercell_end,
             command=command,  # subcommands["command_supercell"]
-            directory=directory / f"{cls.endpoint_relaxation_workflow.name_full}.end",
+            directory=directory / f"{cls.endpoint_energy_workflow.name_full}.end",
             is_restart=is_restart,
         )
+        supercell_start = start_energy.result()
+        supercell_end = end_energy.result()
 
-        # wait for the endpoint relaxations to finish
-        endpoint_start_result = endpoint_start_state.result()
-        endpoint_end_result = endpoint_end_state.result()
-
+        # use the endpoints to generate intermediate images
         images = get_migration_images_from_endpoints(
-            supercell_start=endpoint_start_result,
-            supercell_end=endpoint_end_result,
+            supercell_start=supercell_start,
+            supercell_end=supercell_end,
             nimages=nimages,
         )