+++ docs/change_log.md
@@ -66,10 +66,18 @@ There is one key exception to the rules above -- and that is with `MAJOR`=0 rele
 **Enhancements**
 **Refactors**
 **Fixes**
--->
 
 - no new changes have been merged into the `main` branch yet
+-->
 
+**Enhancements**
+
+- add `relax_bulk` and `relax_endpoints` parameters to optionally turn off pre-relaxations in NEB
+- add CLEASE app for cluster expanison calculations (these workflows are highly experimental at the moment - so use with caution)
+
+**Fixes**
+
+- fix site ordering in NEB supercell structures
 
 --------------------------------------------------------------------------------